+++ models/sam3/sam3/agent/helpers/boxes.py
@@ -129,7 +129,7 @@ class BoxMode(IntEnum):
         if single_box:
             return original_type(arr.flatten().tolist())
         if is_numpy:
-            return arr.numpy()
+            return arr.float().numpy()
         else:
             return arr
