+++ models/sam3/sam3/visualization_utils.py
@@ -47,7 +47,7 @@ def show_img_tensor(img_batch, vis_img_idx=0):
     STD_IMG = np.array([0.5, 0.5, 0.5])
     im_tensor = img_batch[vis_img_idx].detach().cpu()
     assert im_tensor.dim() == 3
-    im_tensor = im_tensor.numpy().transpose((1, 2, 0))
+    im_tensor = im_tensor.float().numpy().transpose((1, 2, 0))
     im_tensor = (im_tensor * STD_IMG) + MEAN_IMG
     im_tensor = np.clip(im_tensor, 0, 1)
     plt.imshow(im_tensor)
@@ -353,7 +353,7 @@ def visualize_formatted_frame_output(
 
                 # Convert back to numpy for plotting
                 mask_np = (
-                    binary_mask.numpy()
+                    binary_mask.float().numpy()
                     if isinstance(binary_mask, torch.Tensor)
                     else binary_mask
                 )
